Move functions to app.tsx + Update Sidebar component
